(refactor) add guard clauses for all migrations

+++ database/migrations/2025_10_03_140740_create_timetable_professors_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('timetable_professors')) {
+            return;
+        }
         Schema::create('timetable_professors', function (Blueprint $table) {
             $table->id();
             $table->foreignId('professor_id')->constrained('professors')->onDelete('cascade');
